# Move time.Now call into exemplar reservoir to improve performance