--- a/sdk/metric/internal/exemplar/drop.go
+++ b/sdk/metric/internal/exemplar/drop.go
@@ -5,20 +5,19 @@ package exemplar // import "go.opentelemetry.io/otel/sdk/metric/internal/exempla
 
 import (
 	"context"
-	"time"
 
 	"go.opentelemetry.io/otel/attribute"
 )
 
-// Drop returns a [Reservoir] that drops all measurements it is offered.
-func Drop() Reservoir { return &dropRes{} }
+// Drop returns a [FilteredReservoir] that drops all measurements it is offered.
+func Drop[N int64 | float64]() FilteredReservoir[N] { return &dropRes[N]{} }
 
-type dropRes struct{}
+type dropRes[N int64 | float64] struct{}
 
 // Offer does nothing, all measurements offered will be dropped.
-func (r *dropRes) Offer(context.Context, time.Time, Value, []attribute.KeyValue) {}
+func (r *dropRes[N]) Offer(context.Context, N, []attribute.KeyValue) {}
 
 // Collect resets dest. No exemplars will ever be returned.
-func (r *dropRes) Collect(dest *[]Exemplar) {
+func (r *dropRes[N]) Collect(dest *[]Exemplar) {
 	*dest = (*dest)[:0]
 }