--- a/sdk/metric/internal/exemplar/drop_test.go
+++ b/sdk/metric/internal/exemplar/drop_test.go
@@ -5,14 +5,20 @@ package exemplar
 
 import (
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func TestDrop(t *testing.T) {
-	t.Run("Int64", ReservoirTest[int64](func(int) (Reservoir, int) {
-		return Drop(), 0
-	}))
+	t.Run("Int64", testDropFiltered[int64])
+	t.Run("Float64", testDropFiltered[float64])
+}
 
-	t.Run("Float64", ReservoirTest[float64](func(int) (Reservoir, int) {
-		return Drop(), 0
-	}))
+func testDropFiltered[N int64 | float64](t *testing.T) {
+	r := Drop[N]()
+
+	var dest []Exemplar
+	r.Collect(&dest)
+
+	assert.Len(t, dest, 0, "non-sampled context should not be offered")
 }