--- a/sdk/metric/internal/exemplar/filter.go
+++ b/sdk/metric/internal/exemplar/filter.go
@@ -5,25 +5,25 @@ package exemplar // import "go.opentelemetry.io/otel/sdk/metric/internal/exempla
 
 import (
 	"context"
-	"time"
 
-	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/trace"
 )
 
-// SampledFilter returns a [Reservoir] wrapping r that will only offer measurements
-// to r if the passed context associated with the measurement contains a sampled
+// Filter determines if a measurement should be offered.
+//
+// The passed ctx needs to contain any baggage or span that were active
+// when the measurement was made. This information may be used by the
+// Reservoir in making a sampling decision.
+type Filter func(context.Context) bool
+
+// SampledFilter is a [Filter] that will only offer measurements
+// if the passed context associated with the measurement contains a sampled
 // [go.opentelemetry.io/otel/trace.SpanContext].
-func SampledFilter(r Reservoir) Reservoir {
-	return filtered{Reservoir: r}
+func SampledFilter(ctx context.Context) bool {
+	return trace.SpanContextFromContext(ctx).IsSampled()
 }
 
-type filtered struct {
-	Reservoir
-}
-
-func (f filtered) Offer(ctx context.Context, t time.Time, n Value, a []attribute.KeyValue) {
-	if trace.SpanContextFromContext(ctx).IsSampled() {
-		f.Reservoir.Offer(ctx, t, n, a)
-	}
+// AlwaysOnFilter is a [Filter] that always offers measurements.
+func AlwaysOnFilter(ctx context.Context) bool {
+	return true
 }