… sdk/metric/internal/aggregate/lastvalue_test.go
… sdk/metric/internal/aggregate/sum_test.go