@MrAlias 2024-06-25
The timestamp arg is something recommended by specification: https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/metrics/sdk.md#exemplarreservoir

This may now be as necessary now, but when we accept a user defined reservoir this seems like an important parameter.

Would passing a zero value time from the SDK and check for this zero value in the reservoir to then call now be a better approach?