@dashpole 2024-06-25
I changed it to pass time.Zero.