@dashpole 2024-06-26
@MrAlias I was re-reading the specification, and I think the intent is that Offer is only called when the exemplar is _definitely_ going to be recorded (i.e. after the filter).  We should probably have one Offer to the (non-public) FilteredReservoir which does not have a timestamp argument, and one Offer to the (to-be-public) Reservoir which does have the timestamp argument.