@MrAlias 2024-06-26
> We should probably have one Offer to the (non-public) FilteredReservoir which does not have a timestamp argument, and one Offer to the (to-be-public) Reservoir which does have the timestamp argument.

I'm not sure I fully understand this implementation, but in general this sounds correct to me.

Does that mean we would keep the existing `Offer` method and just change the `FilteredReservoir` in this PR?