@dashpole 2024-06-26
> Does that mean we would keep the existing Offer method and just change the FilteredReservoir in this PR?

Yes.