@dashpole 2024-06-28
Please let me know if this is too large to review or too large for a last-minute change before a release.  I can try to make a version which just fixes the issue with as small of a diff as possible.